Ensure we retain the original weak key.
The "outer" hash only keys on the bottom 16 bits of the weak key; store the full weak key in the mapped value so we can retrieve it later.

None of this is strictly necessary, it just ensures that this class conforms well to the Map contract. And maybe it still doesn't...